example works but figure is compressed
